chore(node): merge `prt-runner` and `epoch-manager` into one module
