+++ cartesi-rollups/node/dave-rollups/src/main.rs
@@ -1,6 +1,6 @@
 use dave_rollups::{
-    DaveParameters, create_blockchain_reader_task, create_compute_runner_task,
-    create_epoch_manager_task, create_machine_runner_task, create_provider,
+    DaveParameters, create_blockchain_reader_task, create_epoch_manager_task,
+    create_machine_runner_task, create_provider,
 };
 use rollups_state_manager::persistent_state_access::PersistentStateAccess;
 
@@ -24,20 +24,14 @@ async fn main() -> Result<()> {
 
     let blockchain_reader_task =
         create_blockchain_reader_task(state_manager.clone(), provider.clone(), &parameters);
-
     let epoch_manager_task =
         create_epoch_manager_task(provider.clone(), state_manager.clone(), &parameters);
-
     let machine_runner_task = create_machine_runner_task(state_manager.clone(), &parameters);
-
-    let compute_runner_task =
-        create_compute_runner_task(state_manager.clone(), provider.clone(), &parameters);
 
     futures::try_join!(
         blockchain_reader_task,
         epoch_manager_task,
         machine_runner_task,
-        compute_runner_task
     )?;
 
     unreachable!("node services should run forever")
