--- cartesi-rollups/node/prt-runner/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "rollups-prt-runner"
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-edition.workspace = true
-homepage.workspace = true
-license-file.workspace = true
-readme.workspace = true
-repository.workspace = true
-
-[dependencies]
-cartesi-prt-core = { workspace = true }
-rollups-state-manager = { workspace = true }
-
-alloy = { workspace = true }
-log = { workspace = true }
-thiserror = { workspace = true }
