+++ prt/tests/rollups/dave/node.lua
@@ -5,9 +5,9 @@ local function start_dave_node(machine_path, db_path, consensus, input_box, slee
         [[echo $$ ; exec env MACHINE_PATH='%s' STATE_DIR='%s' \
         CONSENSUS='%s' INPUT_BOX='%s' \
         SLEEP_DURATION=%d RUST_BACKTRACE='%s' \
-        RUST_LOG='none',cartesi_prt_core='%s',rollups_prt_runner='%s',rollups_epoch_manager='%s' \
+        RUST_LOG='none',cartesi_prt_core='%s',rollups_epoch_manager='%s' \
         ../../../target/debug/dave-rollups > dave.log 2>&1]],
-        machine_path, db_path, consensus, input_box, sleep_duration, trace_level, verbosity, verbosity, verbosity
+        machine_path, db_path, consensus, input_box, sleep_duration, trace_level, verbosity, verbosity
     )
 
     local reader = io.popen(cmd)
